move authmodal + some minor refactoring + fix some tests
