+++ frontend/package.json
@@ -5,8 +5,8 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "test:unit:coverage": "vue-cli-service test:unit --coverage",
+    "test:unit": "NODE_ENV=test vue-cli-service test:unit --verbose --silent",
+    "test:unit:coverage": "NODE_ENV=test vue-cli-service test:unit --silent --coverage",
     "test:e2e": "vue-cli-service test:e2e",
     "lint": "vue-cli-service lint"
   },
