+++ frontend/src/Shared/components/AuthModal.vue
