+++ frontend/tests/unit/components/addUserModal.spec.js
@@ -16,6 +16,11 @@ jest.mock('axios', () => ({
 }));
 
 describe('AddUserModal Component', () => {
+  beforeAll(async () => {
+    // After this line, router is ready
+    await router.isReady();
+  });
+
   // shallow mount the component
   const wrapper = shallowMount(AddUserModalComponent, {
     props: {
